♻️ refactor: move User entity to auth module and update references
- Move User entity from entities/user.entity.ts to modules/auth/auth.entity.ts
- Extract UserRole enum to schemas package for reuse
- Update import references to User entity across all files
- Fix db:fresh command with explicit config path
- Add db:fresh script to root package.json and turbo.json
- Update entity paths in mikro-orm configuration
