+++ apps/api/package.json
@@ -13,7 +13,7 @@
     "start:prod": "node dist/main",
     "db:create": "mikro-orm database:create && pnpm db:migration:up",
     "db:sync": "mikro-orm schema:update --run",
-    "db:fresh": "mikro-orm schema:fresh --seed --run",
+    "db:fresh": "mikro-orm schema:fresh --seed --run --config=./src/config/mikro-orm.config.ts",
     "db:migration:up": "mikro-orm migration:up",
     "db:migration:down": "mikro-orm migration:down",
     "db:migration:create": "mikro-orm migration:create",
