+++ apps/api/src/entities/index.ts
@@ -10,7 +10,7 @@ export * from './test.entity';
 export * from './workout.entity';
 export * from './workout-category.entity';
 export * from './workout-element.entity';
-export * from './user.entity';
+// export * from './user.entity';
 export * from './media.entity';
 export * from './athlete.entity';
 export * from './club.entity';
