+++ apps/api/src/modules/athlete/use-cases/create-athlete.use-case.ts
@@ -1,10 +1,8 @@
 import { AthleteDto, CreateAthlete } from '@dropit/schemas';
 import { EntityManager } from '@mikro-orm/core';
 import { Injectable } from '@nestjs/common';
-import { Club } from '../../../entities/club.entity';
-import { User } from '../../../entities/user.entity';
-import { AthleteRepository } from '../athlete.repository';
 import { AthletePresenter } from '../athlete.presenter';
+import { AthleteRepository } from '../athlete.repository';
 
 @Injectable()
 export class CreateAthleteUseCase {
