+++ apps/api/src/modules/athlete/use-cases/update-athlete.use-case.ts
@@ -2,10 +2,8 @@
 import { AthleteDto, UpdateAthlete } from '@dropit/schemas';
 import { EntityManager } from '@mikro-orm/core';
 import { Injectable } from '@nestjs/common';
-import { Club } from '../../../entities/club.entity';
-import { User } from '../../../entities/user.entity';
-import { AthleteRepository } from '../athlete.repository';
 import { AthletePresenter } from '../athlete.presenter';
+import { AthleteRepository } from '../athlete.repository';
 @Injectable()
 export class UpdateAthleteUseCase {
   constructor(
