+++ apps/api/src/seeders/athlete.seeder.ts
@@ -1,8 +1,8 @@
+import { UserRole } from '@dropit/schemas';
 import { EntityManager } from '@mikro-orm/core';
 import { Athlete } from '../entities/athlete.entity';
 import { Club } from '../entities/club.entity';
-import { User } from '../entities/user.entity';
-import { UserRole } from '../entities/user.entity';
+import { User } from '../modules/auth/auth.entity';
 
 // Noms et prénoms pour la génération d'athlètes
 const femaleFirstNames = [
