+++ package.json
@@ -15,7 +15,8 @@
     "test:api": "dotenv -e .env -- pnpm --filter api test",
     "test:api:e2e": "dotenv -e .env -- pnpm --filter api test:e2e",
     "test:api:docker": "dotenv -e .env -- pnpm --filter api test:e2e:docker",
-    "db:seed": "dotenv -e .env -- turbo db:seed"
+    "db:seed": "dotenv -e .env -- turbo db:seed",
+    "db:fresh": "dotenv -e .env -- turbo db:fresh"
   },
   "devDependencies": {
     "@biomejs/biome": "1.5.3",
