+++ packages/schemas/src/auth.schema.ts
@@ -1,0 +1,5 @@
+export enum UserRole {
+  ATHLETE = 'athlete',
+  COACH = 'coach',
+  ADMIN = 'admin',
+}
