+++ packages/schemas/src/index.ts
@@ -9,3 +9,4 @@ export * from './workout-category.schema';
 export * from './workout.schema';
 export * from './competitor-status.schema';
 export * from './personal-record.schema';
+export * from './auth.schema';
