+++ turbo.json
@@ -39,6 +39,9 @@
     "db:seed": {
       "cache": false
     },
+    "db:fresh": {
+      "cache": false
+    },
     "clean": {
       "cache": false
     }
